requirements.txt
Flask
flask-jwt-extended
flask-cors
python-dotenv
neo4j
werkzeug
